lambda-tools/requirements.txt

boto
    
